TST: Merge file tests into one file
